Experimental Bank Repacking
timpani_bank files can be repacked with new sound files which have the same name as the original (File replacement basically). Still in experimentation as unindentified bytes are missing. In addition, the timpani_bank that was created is of a smaller size than the original.

The table of contents is also not in the same order as the original; however, as their names are probably referenced in the bank by the bitsquid engine as well as the offsets/lengths of the particular sound file, this should not prove an issue. (I could be horribly wrong though xD).